Suppress conversion warning
Perhaps there are more instances, this one reported by user https://stackoverflow.com/q/78298063/85371

+++ include/boost/core/detail/string_view.hpp
@@ -681,7 +681,7 @@ public:
 
         Ch const* r = traits_type::find( data() + pos, size() - pos, c );
 
-        return r? r - data(): npos;
+        return r? static_cast<size_t>(r - data()): npos;
     }
 
     BOOST_CXX14_CONSTEXPR size_type find( Ch const* s, size_type pos, size_type n ) const BOOST_NOEXCEPT
